fix bug

+++ modules/conways_life/src/conways_life.cpp
@@ -36,7 +36,7 @@ void ConwaysLife::setGrid(const std::vector<std::vector<bool>> &_grid) {
     if (_grid.size() < 3) {
         throw "Error! Height less three";
     }
-    for (int i = 1; i < _grid.size(); i++) {
+    for (int i = 1; i < (int)_grid.size(); i++) {
         if (_grid[0].size() != _grid[i].size()) {
             throw "Error! Grid is not rectangle";
         }
